projects/image_classification_api/requirements.txt
Flask==1.1.4
tensorflow==2.7.0
numpy==1.21.5
Pillow==8.4.0
markupsafe==1.1.1
protobuf==3.20.0
